refactor the code upon comments from reviewers
